requirements.txt
Django==2.0
djangorestframework==3.7.7
python-dotenv==0.7.1
mysqlclient==1.3.12
gunicorn==19.7.1
whitenoise==3.3.1
django-cors-headers==2.1.0
factory-boy==2.9.2
